requirements.txt
streamlit
ollama
Pillow